Unificacion de clases de entidades con la version de Nahu - Eloy
