+++ HelloWorld-Box2d/core/src/com/helloworld/box2dprueba/objetos/Bateria.java
@@ -1,0 +1,23 @@
+package com.helloworld.box2dprueba.objetos;
+
+import com.badlogic.gdx.physics.box2d.World;
+
+public class Bateria extends ItemEquipable {
+
+    int energy;
+
+    public Bateria(World world, int x, int y, int width, int height, boolean isStatic, boolean fixRotation) {
+        super(world, x, y, width, height, isStatic, fixRotation);
+    }
+
+    //setter & getter
+
+    public int getEnergy() {
+        return energy;
+    }
+
+    public void setEnergy(int energy) {
+        this.energy = energy;
+    }
+}
+
