+++ HelloWorld-Box2d/core/src/com/helloworld/box2dprueba/objetos/ItemEquipable.java
@@ -1,4 +1,25 @@
 package com.helloworld.box2dprueba.objetos;
 
-public abstract class ItemEquipable {
+import com.badlogic.gdx.physics.box2d.World;
+import com.helloworld.box2dprueba.entidades.Entidad;
+
+public abstract class ItemEquipable extends Entidad {
+
+    private boolean estaEquipado;
+
+    //Constructor
+    public ItemEquipable(World world, int x, int y, int width, int height, boolean isStatic, boolean fixRotation) {
+        super(world, x, y, width, height, isStatic, fixRotation);
+        this.estaEquipado = false;
+    }
+
+
+    //Getter & Setter
+    public boolean isEstaEquipado() {
+        return estaEquipado;
+    }
+
+    public void setEstaEquipado(boolean estaEquipado) {
+        this.estaEquipado = estaEquipado;
+    }
 }
