+++ HelloWorld-Box2d/core/src/com/helloworld/box2dprueba/objetos/Llave.java
@@ -1,8 +1,8 @@
-package com.helloworld.box2dprueba.entidades;
+package com.helloworld.box2dprueba.objetos;
 
 import com.badlogic.gdx.physics.box2d.World;
 
-public class Llave extends ItemEquipable{
+public class Llave extends ItemEquipable {
 
     //Constructor
     public Llave(World mundo, float posX, float posY, float largo, float ancho, String textura, boolean estaEquipado) {
