--- HelloWorld-Box2d/core/src/com/helloworld/box2dprueba/objetos/Luminaria.java
@@ -1,9 +1,0 @@
-package com.helloworld.box2dprueba.objetos;
-
-public abstract class Luminaria {
-
-
-
-    public void dispose(){}
-
-}
